Expanded IStory, added SampleGameXML

+++ app/src/main/kotlin/com/github/paulschaaf/gargoyle/model/IStory.kt
@@ -39,5 +39,5 @@ interface IStory {
   val ratingCountAvg: Int?
   val ratingCountTotal: Int?
   val seriesNumber: Int?
-  val starRating: Int?
+  val starRating: Double?
 }
